src/feedback_backend/requirements.txt
flask
psycopg2-binary
flask_cors
prometheus_flask_exporter
flask_oidc
PyJWT>=2.0.0
flask-cors
requests
python-keycloak
python-jose

